+++ src/logger/index.js
@@ -2,8 +2,9 @@ import winston from 'winston'
 import moment from 'moment-timezone'
 import chalk from 'chalk'
 import config from 'config'
+import {dbs, publish, push} from '../redis/db'
 
-const COLORS = {
+export const COLORS = {
   gray: 'gray',
   redBright: 'redBright',
   greenBright : 'greenBright',
@@ -29,41 +30,64 @@ const alignColorsAndTime = winston.format.combine(
     /*eslint-enable */
 )
 
-const logger = winston.createLogger({
-  level: "info",
-  transports: [
-    new winston.transports.File({ filename: 'error.log', level: 'error' }),
-    new winston.transports.File({ filename: 'combined.log' }),
-      new (winston.transports.Console)({
-          format: winston.format.combine(winston.format.colorize(), alignColorsAndTime)
-      })
-  ],
-})
+function createTransport({filename, level}) {
+  return new winston.transports.File({filename, level})
+}
+
+function createLogger(transports) {
+  return winston.createLogger({
+    level: "info",
+    transports: (transports ? transports.map(createTransport) : [
+      new winston.transports.File({ filename: 'error.log', level: 'error' }),
+      new winston.transports.File({ filename: 'info.log', level: 'info' }),
+    ]).concat(new (winston.transports.Console)({
+      format: winston.format.combine(winston.format.colorize(), alignColorsAndTime)
+      })),
+  })
+}
+let logger
 
 export class Logger {
-  constructor({appName, timezone, hostIp, hostname}) {
+  constructor({appName, timezone, hostIp, hostname, errorIgnoreLevels = [], transports}, redisConnections = {}) {
+    logger = createLogger(transports)
     this.env = process.NODE_ENV || 'dev',
     this.appName = appName
     this.timezone = timezone
     this.hostIp = process.env.HOST_IP || hostIp
     this.hostname = process.env.HOSTNAME || hostname
+    this.errorIgnoreLevels = errorIgnoreLevels
+    const {publisher, client} = redisConnections
+    if (publisher) {
+      dbs[publisher] = publisher
+      this.publisher = publisher
+    }
+    if (client) {
+      dbs[client] = client
+      this.client = client
+    }
+  }
+  getInoreLevels() {
+    return this.errorIgnoreLevels
   }
   log (message = '', metadata = {}) {
-    logger
-      .log({
-      env: this.env,
-      hostIp: this.hostIp,
-      hostname: this.hostname,
-      appName: this.appName,
-      timezone: this.timezone,
-      level: 'info',
-      message,
-      metadata})
+    const payload = {
+    env: this.env,
+    hostIp: this.hostIp,
+    hostname: this.hostname,
+    appName: this.appName,
+    timezone: this.timezone,
+    level: 'info',
+    message,
+    metadata}
+    logger.log(payload)
+    if (this.publisher)
+      publish(`${this.env}:${this.appName}`, payload)
+    if (this.client)
+      push(`${this.env}:${this.appName}`, payload)
   }
   error (message= '', metadata = {}) {
-    logger
-      .error({
-      env: process.NODE_ENV || 'dev',
+    const payload = {
+      env: this.env,
       hostIp: this.hostIp,
       hostname: this.hostname,
       appName: this.appName,
@@ -71,7 +95,15 @@ export class Logger {
       level: 'error',
       message,
       metadata
-      })
+    }
+    logger
+      .error(payload)
+    if (this.publisher)
+      publish(`${this.env}:${this.appName}`, payload)
+    if (this.client)
+      push(`${this.env}:${this.appName}`, payload)
   }
 }
 export default new Logger(config.get('env'))
+
+export const getLogger = (config, redisConnections) => new Logger(config, redisConnections)