+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [.0.0.3] - 2019-03-26
+##### [.0.0.3] - 2019-03-26
 - Moved Logger from server to middleware
 - Changed instantiation of Logger to an instance from a static class
 - Finished Test runner script
@@ -12,8 +12,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed an issue with the redis_db pop signature
 
 
-## [.0.0.5] - 2019-03-26
+##### [.0.0.5] - 2019-03-26
 - Fixed an issue where the range command was not giving the results correctly
 - Fixed some documentation typos
 - Updated roadmap to account for bug fix release
 - Removed extraneous async methods
+
+##### [.0.0.6] - 2019-04-26
+- Added publisher/subscribe routes and scripts
+- Added documentation for publish/subscribe
+- Added tests for publish functions
+- Fixed some documentation typos
+- Changed roadmap estimations based on availability